Refactor discambwrapper into multiple files
